remover casa de um dado fornecedor e adicao de smartcameras a funcionar

+++ SmartHouses/src/Model/SmartCamera.java
@@ -4,13 +4,13 @@ public class SmartCamera extends SmartDevice{
     private String resolucao;//(z x y)
     private double tamanho;//em MB
 
-    public SmartCamera(String id, boolean modo, String res,int tamanho,double consumoBase){
+    public SmartCamera(String id, boolean modo, String res,double tamanho,double consumoBase){
         super(id,modo,consumoBase);
         this.tamanho=tamanho;
         this.resolucao=res;
     }
 
-    public SmartCamera(String id, String res,int tamanho,double consumoBase){
+    public SmartCamera(String id, String res,double tamanho,double consumoBase){
         super(id,consumoBase);
         this.tamanho=tamanho;
         this.resolucao=res;
@@ -56,7 +56,7 @@ public class SmartCamera extends SmartDevice{
 
     public static SmartCamera parseSmartCamera(String line){
         String[] parte = line.split(",");
-        return new SmartCamera(parte[3],parte[0],Integer.parseInt(parte[1]),Double.parseDouble(parte[2]));
+        return new SmartCamera(parte[3],parte[0],Double.parseDouble(parte[1]),Double.parseDouble(parte[2]));
     }
     //    public Model.SmartCamera(String id, boolean modo, LocalDateTime on, LocalDateTime off,int res,double tamanho){
     //    Model.SmartCamera:(1024x768),67,4.00,cam2,2022-03-14 10:53:07,2022-05-11 07:16:25,true
